tree view: format preamble consistently with blob view
Wrap the ref in paragraph tags so the vertical alignment is consistent
with the content in the blob.html template, so that when the user
clicks through from a directory listing to an individual file the
header section of the page does not appear to change sizes.

Remove the extra path value, since it is the same as the parent path.

Signed-off-by: Doug Hellmann <doug@doughellmann.com>