patch view: introduce a view for commits ending in .patch
This new view introduces a text/plain version of a single patch. It has
some differences when using `git format-patch` and will match more the
`--no-stats` version of the refered git command. This is due the fact
that go-git has no support for summary for diffs yet (I have a patch
waiting for over a month there) [1].

There's also a problem when dealing with the root commit, which go-git
Patch API still doesn't support [2].

[1]: https://github.com/go-git/go-git/pull/301
[2]: https://github.com/go-git/go-git/issues/281